% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_attr}
\alias{get_attr}
\title{Extract attribute value from XML node attributes}
\usage{
get_attr(node, attr_name, default = NULL)
}
\arguments{
\item{node}{XML node}

\item{attr_name}{Name of attribute to extract}

\item{default}{Default value if attribute not found}
}
\value{
Attribute value or default
}
\description{
Extract attribute value from XML node attributes
}
