% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_dict}
\alias{parse_dict}
\title{Parse dictionary/list recursively}
\usage{
parse_dict(obj)
}
\arguments{
\item{obj}{List object potentially containing HTML content}
}
\value{
List with parsed HTML content
}
\description{
Parse dictionary/list recursively
}
