% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{raw_to_int}
\alias{raw_to_int}
\title{Read SnapGene files and convert to R list structure
Required packages: xml2, httr, stringr, Biostrings, IRanges, bit64
Convert raw bytes to integer (big endian)}
\usage{
raw_to_int(bytes)
}
\arguments{
\item{bytes}{Raw vector of bytes}
}
\value{
Integer value
}
\description{
Read SnapGene files and convert to R list structure
Required packages: xml2, httr, stringr, Biostrings, IRanges, bit64
Convert raw bytes to integer (big endian)
}
