% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_snapgene.R
\name{read_snapgene}
\alias{read_snapgene}
\title{Convert SnapGene file to R list}
\usage{
read_snapgene(filepath)
}
\arguments{
\item{filepath}{Path to SnapGene .dna file}
}
\value{
List containing parsed SnapGene data
}
\description{
Convert SnapGene file to R list
}
