% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapgeneToDnaStringset.R
\name{snapgene_file_to_dnastringset}
\alias{snapgene_file_to_dnastringset}
\title{Convert SnapGene file to Biostrings DNAStringSet}
\usage{
snapgene_file_to_dnastringset(filepath)
}
\arguments{
\item{filepath}{Path to SnapGene .dna file}
}
\value{
Biostrings DNAStringSet object with features and annotations
}
\description{
Convert SnapGene file to Biostrings DNAStringSet
}
